use int32 for count

+++ backend/src/apiserver/server/experiment_server.go
@@ -57,7 +57,7 @@ func (s *ExperimentServer) ListExperiment(ctx context.Context, request *api.List
 	}
 	return &api.ListExperimentsResponse{
 			Experiments:   ToApiExperiments(experiments),
-			Count:         int32(count),
+			Count:         count,
 			NextPageToken: nextPageToken},
 		nil
 }
