+++ backend/src/apiserver/server/pipeline_server.go
@@ -89,7 +89,7 @@ func (s *PipelineServer) ListPipelines(ctx context.Context, request *api.ListPip
 		return nil, util.Wrap(err, "List pipelines failed.")
 	}
 	apiPipelines := ToApiPipelines(pipelines)
-	return &api.ListPipelinesResponse{Pipelines: apiPipelines, Count: int32(count), NextPageToken: nextPageToken}, nil
+	return &api.ListPipelinesResponse{Pipelines: apiPipelines, Count: count, NextPageToken: nextPageToken}, nil
 }
 
 func (s *PipelineServer) DeletePipeline(ctx context.Context, request *api.DeletePipelineRequest) (*empty.Empty, error) {
